requirements.txt
numpy==1.23.4
sklearn==0.0
Pillow==9.3.0
scipy==1.9.3
pefile
enum34
joblib
flask==2.1.3
flask-restx==0.5.1
werkzeug==2.1.2
sqlitedict
jsonlines
python-dotenv
